Rename the 'stringifiers' subpackage to the 'convert' and move to the 'org.spine3' package
