+++ client/src/main/java/org/spine3/validate/Validate.java
@@ -31,8 +31,8 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static java.lang.String.format;
-import static org.spine3.base.stringifiers.Stringifiers.EMPTY_ID;
-import static org.spine3.base.stringifiers.Stringifiers.idToString;
+import static org.spine3.convert.Stringifiers.EMPTY_ID;
+import static org.spine3.convert.Stringifiers.idToString;
 
 /**
  * This class provides general validation routines.
