+++ client/src/test/java/org/spine3/base/IdentifiersShould.java
@@ -22,14 +22,13 @@ package org.spine3.base;
 
 import com.google.common.testing.NullPointerTester;
 import org.junit.Test;
-import org.spine3.base.Identifiers;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
 import static org.spine3.base.Identifiers.idToAny;
 import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.base.stringifiers.Stringifiers.idToString;
+import static org.spine3.convert.Stringifiers.idToString;
 import static org.spine3.test.Tests.assertHasPrivateParameterlessCtor;
 
 /**
