+++ server/src/main/java/org/spine3/server/entity/AbstractVersionableEntity.java
@@ -33,7 +33,7 @@ import java.util.Objects;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static java.lang.String.format;
-import static org.spine3.base.stringifiers.Stringifiers.idToString;
+import static org.spine3.convert.Stringifiers.idToString;
 import static org.spine3.base.Versions.checkIsIncrement;
 
 /**
