+++ server/src/main/java/org/spine3/server/validate/AbstractValidatingBuilder.java
@@ -24,9 +24,9 @@ import com.google.common.base.Optional;
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Message;
 import org.spine3.base.FieldPath;
-import org.spine3.base.stringifiers.RegistryKey;
-import org.spine3.base.stringifiers.Stringifier;
-import org.spine3.base.stringifiers.StringifierRegistry;
+import org.spine3.convert.RegistryKey;
+import org.spine3.convert.Stringifier;
+import org.spine3.convert.StringifierRegistry;
 import org.spine3.validate.ConstraintViolation;
 import org.spine3.validate.ConstraintViolationThrowable;
 import org.spine3.validate.ConversionError;
